Get minimal assessment data for a Latin binomial

The assessment_by_ functions including assessments_by_name() return 'minimal' assessment data for your specified filters (arguments). The minimal assessment data provides assessment_ids and sis_taxon_ids (as a tibble) which can be used with assessment_data(), assessment_data_many() and parse_assessment_data() to get full assessment data.

The assessments_by_name() function returns minimal assessment data for a supplied Latin binomial. The genus and species arguments are not case sensitive.

Usage

assessments_by_name(api, genus, species)

Arguments

api

An httr2 response object created with init_api().

genus

String. A genus name.

species

String. A species name.

Value

Returns a tibble() of minimal assessment data for the supplied genus and species name. The minimal assessment data provides assessment_ids and sis_taxon_ids which can then be used with assessment_data(), assessment_data_many() and parse_assessment_data() to get full assessment data.

Examples

if (FALSE) { # \dontrun{
assessments_by_name(api, "panthera", "leo")
} # }