Extract data from a named element

This is a helper function to return a tidy tibble() for a specified element name. Here, an 'element name' is one of the named tibbles created when running parse_assessement_data() or assessment_data_many().

To get a list of valid element names that you can concatenate with extract_element() first call element_names(). Pass one of these names to extract_element() to create a tidy tibble() of that specific data.

Usage

extract_element(assessments_list, element_name)

Arguments

assessments_list

A list containing full assessment data.

element_name

String.

Value

Returns a tidy tibble() of data for the specified element name.

Examples

if (FALSE) { # \dontrun{
extract_element(assessments_list, "habitats")
extract_element(assessments_list, "biogeographical_realms")
extract_element(assessments_list, "taxon_synonyms")
extract_element(assessments_list, "threats")
} # }